Get register and login forms working
Update registration form and add login form, using the React router  Form component and actions to handle updating and submitting form data.
Make sure the API responds with an HTTP status on success for both registration and login rather than a redirect.
Update DB so all fields match what we're collecting on the front end. Switch from user name to email in case we want to send users emails in the future.
Set up CORS handling in the API so it can accept requests from the front end app in development.